+++ Classes/AStarNode.h
@@ -7,18 +7,18 @@ class AStarNode
 public:
 	int x;
 	int y;
-	int f;
-	int g;
-	int h;
+	float f;
+	float g;
+	float h;
 	int px;
 	int py;
 
 	AStarNode(void);
 	~AStarNode(void);
 	void setPos(int,int);
-	void setF(int);
-	void setG(int);
-	void setH(int);
+	void setF(float);
+	void setG(float);
+	void setH(float);
 	//CREATE_FUNC(AStarNode);
 	bool operator < (const AStarNode& n) const;
 	bool operator == (const AStarNode& n) const;
